+++ constants.js
@@ -20,4 +20,9 @@ export const intervalIdsMap = {
   5: 'months',
 };
 
+export const renewFromMap = {
+  CURRENT_DUE_DATE: '1',
+  SYSTEM_DATE: '2',
+};
+
 export default {};