+++ withRenew.js
@@ -1,10 +1,10 @@
-import _ from 'lodash';
+import { omit, get } from 'lodash';
 import moment from 'moment'; // eslint-disable-line import/no-extraneous-dependencies
 import React from 'react';
 import PropTypes from 'prop-types';
 
 import ErrorModal from './lib/ErrorModal';
-import { getFixedDueDateSchedule, calculateDueDate } from './util';
+import { getFixedDueDateSchedule, calculateDueDate, isLoanProfileFixed } from './loanUtils';
 
 // HOC used to manage renew
 const withRenew = WrappedComponent =>
@@ -72,8 +72,9 @@ const withRenew = WrappedComponent =>
 
     renew(data) {
       return this.fetchLoanPolicy(data)
-        .then(loan => this.fetchFixedDueDateSchedules(loan))
-        .then(loan => this.validateFixedDueSchedule(loan))
+        .then(loan => this.fetchFixedDueDateSchedule(loan))
+        .then(loan => this.fetchAlternateFixedDueDateSchedule(loan))
+        .then(loan => this.validateRenew(loan))
         .then(loan => this.execRenew(loan))
         .catch((error) => {
           this.setState({ error });
@@ -91,61 +92,96 @@ const withRenew = WrappedComponent =>
       });
     }
 
-    fetchFixedDueDateSchedules(loan) {
-      if (!loan || !loan.loanPolicy || !loan.loanPolicy.loansPolicy.fixedDueDateSchedule) {
-        return loan;
-      }
+    fetchFixedDueDateSchedule(loan) {
+      const scheduleId = get(loan, 'loanPolicy.loansPolicy.fixedDueDateSchedule', '');
+      if (!scheduleId) return loan;
 
-      const query = `(id=="${loan.loanPolicy.loansPolicy.fixedDueDateSchedule}")`;
-      this.props.mutator.fixedDueDateSchedules.reset();
-      return this.props.mutator.fixedDueDateSchedules.GET({ params: { query } }).then((fixedDueDateSchedules) => {
-        loan.loanPolicy.fixedDueDateSchedule = fixedDueDateSchedules[0];
+      return this.fetchSchedule(loan, scheduleId).then((schedule) => {
+        loan.loanPolicy.fixedDueDateSchedule = schedule;
         return loan;
       });
     }
 
-    // eslint-disable-next-line class-methods-use-this
-    validateFixedDueSchedule(loan) {
+    fetchAlternateFixedDueDateSchedule(loan) {
+      const scheduleId = get(loan, 'loanPolicy.renewalsPolicy.alternateFixedDueDateScheduleId', '');
+      if (!scheduleId) return loan;
+
+      return this.fetchSchedule(loan, scheduleId).then((schedule) => {
+        loan.loanPolicy.alternateFixedDueDateSchedule = schedule;
+        return loan;
+      });
+    }
+
+    fetchSchedule(loan, scheduleId) {
+      const query = `(id=="${scheduleId}")`;
+      this.props.mutator.fixedDueDateSchedules.reset();
+      return this.props.mutator.fixedDueDateSchedules.GET({ params: { query } })
+        .then(fixedDueDateSchedules => fixedDueDateSchedules[0]);
+    }
+
+    validateRenew(loan) {
       const { loanPolicy } = loan;
-      if (loanPolicy && loanPolicy.fixedDueDateSchedule) {
-        const schedule = getFixedDueDateSchedule(loanPolicy.fixedDueDateSchedule.schedules);
+      const loansPolicy = loanPolicy.loansPolicy || {};
 
-        if (!schedule) {
-          const error = {
-            message: ` Item can't be renewed as the renewal date falls outside
-              of the date ranges in the loan policy.
-              Please review ${loan.loanPolicy.name} before retrying renewal.`,
-          };
-
-          return Promise.reject(error);
-        }
-
-        loanPolicy.fixedDueDateSchedule.schedule = schedule;
+      if (isLoanProfileFixed(loansPolicy)) {
+        this.validateSchedules(loanPolicy);
       }
 
-      const newDueDate = calculateDueDate(loan);
+      this.validateDueDate(loan);
+
+      return loan;
+    }
+
+    validateSchedules(loanPolicy) {
+      const { fixedDueDateSchedule, alternateFixedDueDateSchedule } = loanPolicy;
+      const schedule = this.validateSchedule(loanPolicy, alternateFixedDueDateSchedule);
+
+      if (!schedule) {
+        this.validateSchedule(loanPolicy, fixedDueDateSchedule);
+      }
+    }
+
+    validateSchedule(loanPolicy, dueDateSchedule) {
+      if (!dueDateSchedule) return dueDateSchedule;
+
+      const schedule = getFixedDueDateSchedule(dueDateSchedule.schedules);
+
+      if (!schedule) {
+        return this.throwError(` Item can't be renewed as the renewal date falls outside
+          of the date ranges in the loan policy.
+          Please review ${loanPolicy.name} before retrying renewal.`);
+      }
+
+      dueDateSchedule.schedule = schedule;
+      return dueDateSchedule;
+    }
+
+    validateDueDate(loan) {
+      const newDueDate = calculateDueDate(loan).startOf('day');
       const currentDueDate = moment(loan.dueDate).startOf('day');
 
-      if (!currentDueDate.diff(newDueDate, 'days')) {
-        const error = {
-          message: 'Renewal at this time would not change the due date.',
-        };
-
-        return Promise.reject(error);
+      if (newDueDate.isSameOrBefore(currentDueDate)) {
+        return this.throwError('Renewal at this time would not change the due date.');
       }
 
       return loan;
     }
 
+    // eslint-disable-next-line class-methods-use-this
+    throwError(message) {
+      const error = { message };
+      throw error;
+    }
+
     execRenew(loan) {
       Object.assign(loan, {
         renewalCount: (loan.renewalCount || 0) + 1,
-        loanDate: moment(loan.loanDate).format(),
-        dueDate: calculateDueDate(loan),
+        loanDate: moment(loan.loanDate).utc().format(),
+        dueDate: calculateDueDate(loan).utc().format(),
         action: 'renewed',
       });
 
-      const loanData = _.omit(loan, ['item', 'rowIndex', 'loanPolicy']);
+      const loanData = omit(loan, ['item', 'rowIndex', 'loanPolicy']);
 
       this.props.mutator.loanId.replace(loan.id);
       return this.props.mutator.loansHistory.PUT(loanData);